# packages/backend/src/OpenApi/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Documentation",
    "version": "1.0.0"
  },
  "components": {
    "schemas": {
...
      "Election": {
        "properties": {
          "admin_ids": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "audit_ids": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "auth_key": {
            "type": "string"
          },
          "claim_key_hash": {
            "type": "string"
          },
          "create_date": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "string"
              }
            ]
          },
          "credential_ids": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "description": {
            "type": "string"
          },
          "election_id": {
            "type": "string"
          },
          "end_time": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "string"
              }
            ]
          },
          "frontend_url": {
            "type": "string"
          },
          "head": {
            "type": "boolean"
          },
          "is_public": {
            "type": "boolean"
          },
          "owner_id": {
            "type": "string"
          },
          "races": {
            "items": {
              "$ref": "#/components/schemas/Race"
            },
            "type": "array"
          },
          "settings": {
            "$ref": "#/components/schemas/ElectionSettings"
          },
          "start_time": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "string"
              }
            ]
          },
          "state": {
            "$ref": "#/components/schemas/ElectionState"
          },
          "support_email": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "update_date": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "string"
              }
            ]
          }
        },
        "type": "object"
      },
      "ElectionResults": {
        "anyOf": [
          {
            "properties": {
              "results": {
                "$ref": "#/components/schemas/starResults"
              },
              "votingMethod": {
                "const": "STAR",
                "type": "string"
              }
            },
            "type": "object"
          },
          {
            "properties": {
              "results": {
                "$ref": "#/components/schemas/allocatedScoreResults"
              },
              "votingMethod": {
                "const": "STAR_PR",
                "type": "string"
              }
            },
            "type": "object"
          },
          {
            "properties": {
              "results": {
                "$ref": "#/components/schemas/approvalResults"
              },
              "votingMethod": {
                "const": "Approval",
                "type": "string"
              }
            },
            "type": "object"
          },
          {
            "properties": {
              "results": {
                "$ref": "#/components/schemas/rankedRobinResults"
              },
              "votingMethod": {
                "const": "RankedRobin",
                "type": "string"
              }
            },
            "type": "object"
          },
          {
            "properties": {
              "results": {
                "$ref": "#/components/schemas/irvResults"
              },
              "votingMethod": {
                "const": "IRV",
                "type": "string"
              }
            },
            "type": "object"
          },
          {
            "properties": {
              "results": {
                "$ref": "#/components/schemas/pluralityResults"
              },
              "votingMethod": {
                "const": "Plurality",
                "type": "string"
              }
            },
            "type": "object"
          },
          {
            "properties": {
              "results": {
                "$ref": "#/components/schemas/irvResults"
              },
              "votingMethod": {
                "const": "STV",
                "type": "string"
              }
            },
            "type": "object"
          }
        ]
      },
      "ElectionRoll": {
        "properties": {
          "address": {
            "type": "string"
          },
          "ballot_id": {
            "type": "string"
          },
          "create_date": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "string"
              }
            ]
          },
          "election_id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "email_data": {
            "properties": {
              "inviteResponse": {},
              "reminderResponse": {}
            },
            "type": "object"
          },
          "head": {
            "type": "boolean"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/ElectionRollAction"
            },
            "type": "array"
          },
          "ip_hash": {
            "type": "string"
          },
          "precinct": {
            "type": "string"
          },
          "registration": {},
          "state": {
            "$ref": "#/components/schemas/ElectionRollState"
          },
          "submitted": {
            "type": "boolean"
          },
          "update_date": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "string"
              }
            ]
          },
          "voter_id": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "ElectionRollAction": {
        "properties": {
          "action_type": {
            "type": "string"
          },
          "actor": {
            "type": "string"
          },
          "timestamp": {
            "type": "number"
          }
        },
        "type": "object"
      },
      "ElectionRollState": {
        "enum": [
          "approved",
          "flagged",
          "registered",
          "invalid"
        ],
        "type": "string"
      },
      "ElectionSettings": {
        "properties": {
          "ballot_updates": {
            "type": "boolean"
          },
          "break_ties_randomly": {
            "type": "boolean"
          },
          "invitation": {
            "enum": [
              "address",
              "email"
            ],
            "type": "string"
          },
          "max_rankings": {
            "type": "number"
          },
          "public_results": {
            "type": "boolean"
          },
          "random_candidate_order": {
            "type": "boolean"
          },
          "reminders": {
            "type": "boolean"
          },
          "require_instruction_confirmation": {
            "type": "boolean"
          },
          "term_type": {
            "enum": [
              "election",
              "poll"
            ],
            "type": "string"
          },
          "time_zone": {
            "type": "string"
          },
          "voter_access": {
            "enum": [
              "closed",
              "open",
              "registration"
            ],
            "type": "string"
          },
          "voter_authentication": {
            "$ref": "#/components/schemas/authentication"
          }
        },
        "type": "object"
      },
      "ElectionState": {
        "enum": [
          "archived",
          "closed",
          "draft",
          "finalized",
          "open"
        ],
        "type": "string"
      },
...
      "Race": {
        "properties": {
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/Candidate"
            },
            "type": "array"
          },
          "description": {
            "type": "string"
          },
          "num_winners": {
            "type": "number"
          },
          "precincts": {
            "items": {
              "additionalProperties": false,
              "patternProperties": {
                "^[0-9]+$": {
                  "type": "string"
                }
              },
              "properties": {
                "length": {
                  "type": "number"
                }
              },
              "type": "object"
            },
            "type": "array"
          },
          "race_id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "voting_method": {
            "$ref": "#/components/schemas/VotingMethod"
          }
        },
        "type": "object"
      },
...
      "VotingMethod": {
        "enum": [
          "Approval",
          "IRV",
          "Plurality",
          "RankedRobin",
          "STAR",
          "STAR_PR",
          "STV"
        ],
        "type": "string"
      },
...
      "roles": {
        "enum": [
          "owner",
          "admin",
          "auditor",
          "system_admin",
          "credentialer"
        ],
        "type": "string"
      },
...
    }
  },
  "paths": {
...
    "/Election/{id}": {
      "get": {
        "summary": "Get election by ID",
        "tags": [
          "Elections"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The election ID"
          }
        ],
        "responses": {
          "200": {
            "description": "The election description by ID",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Election"
                }
              }
            }
          },
          "404": {
            "description": "Election not found"
          }
        }
      },
      "delete": {
        "summary": "Delete election by ID",
        "tags": [
          "Elections"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The election ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Election deleted"
          },
          "404": {
            "description": "Election not found"
          }
        }
      }
    },
    "/Election/{_id}/exists": {
      "get": {
        "summary": "Check if election exists by ID",
        "tags": [
          "Elections"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The election ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Election exists",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "exists": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Election not found"
          }
        }
      }
    },
...
    "/Election/{id}/edit": {
      "post": {
        "summary": "Edit an election",
        "tags": [
          "Elections"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The election ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Election edited",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "election": {
                      "type": "object",
                      "$ref": "#/components/schemas/Election"
                    },
                    "voterAuth": {
                      "type": "object",
                      "properties": {
                        "authorized_voter": {
                          "type": "boolean"
                        },
                        "has_voted": {
                          "type": "boolean"
                        },
                        "roles": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/roles"
                          }
                        },
                        "permissions": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/permissions"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Election not found"
          }
        }
      }
    },
    "/Election/{id}/roles": {
      "put": {
        "summary": "Edit election roles",
        "tags": [
          "Elections"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The election ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Roles edited",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "election": {
                      "type": "object",
                      "$ref": "#/components/schemas/Election"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Election not found"
          }
        }
      }
    },
    "/ElectionResult/{id}": {
      "get": {
        "summary": "Get election results by ID",
        "tags": [
          "Elections"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The election ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Election results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "election": {
                      "type": "object",
                      "$ref": "#/components/schemas/Election"
                    },
                    "results": {
                      "type": "object",
                      "$ref": "#/components/schemas/ElectionResults"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Election not found"
          }
        }
      }
    },
    "/Election/{id}/finalize": {
      "post": {
        "summary": "Finalize an election",
        "tags": [
          "Elections"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The election ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Election finalized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "election": {
                      "type": "object",
                      "$ref": "#/components/schemas/Election"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Election not found"
          }
        }
      }
    },
    "/Election/{id}/setPublicResults": {
      "post": {
        "summary": "Set public results for an election",
        "tags": [
          "Elections"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The election ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Public results set",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "election": {
                      "type": "object",
                      "$ref": "#/components/schemas/Election"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Election not found"
          }
        }
      }
    },
    "/Election/{id}/archive": {
      "post": {
        "summary": "Archive an election",
        "tags": [
          "Elections"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The election ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Election archived",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "election": {
                      "type": "object",
                      "$ref": "#/components/schemas/Election"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Election not found"
          }
        }
      }
    },
    "/Election/{id}/sendInvites": {
      "post": {
        "summary": "Send invitations for an election",
        "tags": [
          "Elections"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The election ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Invitations sent"
          },
          "404": {
            "description": "Election not found"
          }
        }
      }
    },
    "/Election/{id}/sendInvite/{voter_id}": {
      "post": {
        "summary": "Send an invitation to a specific voter",
        "tags": [
          "Elections"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The election ID"
          },
          {
            "in": "path",
            "name": "voter_id",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "The voter ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Invitation sent",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "electionRollEntry": {
                      "type": "object",
                      "$ref": "#/components/schemas/ElectionRoll"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Election or voter not found"
          }
        }
      }
    },
...
    "/images": {
      "post": {
        "summary": "Upload an image",
        "tags": [
          "Elections"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Image uploaded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "photo_filename": {
                      "typeOf": [
                        "string",
                        null
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}